@@ -6,13 +6,13 @@ typedef PlatformViewCreatedCallback = void Function(int id);
 
 class ArCoreAndroidView extends AndroidView {
   final String viewType;
-  final PlatformViewCreatedCallback onPlatformViewCreated;
+  final PlatformViewCreatedCallback? onPlatformViewCreated;
   final ArCoreViewType arCoreViewType;
   final bool debug;
 
   ArCoreAndroidView(
-      {Key key,
-      @required this.viewType,
+      {Key? key,
+      required this.viewType,
       this.onPlatformViewCreated,
       this.arCoreViewType = ArCoreViewType.STANDARDVIEW,
       this.debug = false})
