@@ -1,26 +1,24 @@
 import 'dart:typed_data';
-
-import 'package:flutter/cupertino.dart';
 import 'package:flutter/services.dart';
 
 import '../arcore_flutter_plugin.dart';
 
 class ArCoreFaceController {
   ArCoreFaceController(
-      {int id, this.enableAugmentedFaces, this.debug = false}) {
+      {int? id, this.enableAugmentedFaces, this.debug = false}) {
     _channel = MethodChannel('arcore_flutter_plugin_$id');
-    _channel.setMethodCallHandler(_handleMethodCalls);
+    _channel!.setMethodCallHandler(_handleMethodCalls);
     init();
   }
 
-  final bool enableAugmentedFaces;
+  final bool? enableAugmentedFaces;
   final bool debug;
-  MethodChannel _channel;
-  StringResultHandler onError;
+  MethodChannel? _channel;
+  StringResultHandler? onError;
 
   init() async {
     try {
-      await _channel.invokeMethod<void>('init', {
+      await _channel!.invokeMethod<void>('init', {
         'enableAugmentedFaces': enableAugmentedFaces,
       });
     } on PlatformException catch (ex) {
@@ -35,7 +33,7 @@ class ArCoreFaceController {
     switch (call.method) {
       case 'onError':
         if (onError != null) {
-          onError(call.arguments);
+          onError!(call.arguments);
         }
         break;
       default:
@@ -47,9 +45,8 @@ class ArCoreFaceController {
   }
 
   Future<void> loadMesh(
-      {@required Uint8List textureBytes, String skin3DModelFilename}) {
-    assert(textureBytes != null);
-    return _channel.invokeMethod('loadMesh', {
+      {required Uint8List textureBytes, String? skin3DModelFilename}) {
+    return _channel!.invokeMethod('loadMesh', {
       'textureBytes': textureBytes,
       'skin3DModelFilename': skin3DModelFilename
     });
