@@ -8,12 +8,12 @@ import 'arcore_view.dart';
 typedef void ArCoreFaceViewCreatedCallback(ArCoreFaceController controller);
 
 class ArCoreFaceView extends StatefulWidget {
-  final ArCoreFaceViewCreatedCallback onArCoreViewCreated;
+  final ArCoreFaceViewCreatedCallback? onArCoreViewCreated;
   final bool enableAugmentedFaces;
   final bool debug;
 
   const ArCoreFaceView(
-      {Key key,
+      {Key? key,
       this.onArCoreViewCreated,
       this.enableAugmentedFaces = false,
       this.debug = false})
@@ -27,7 +27,7 @@ class _ArCoreFaceViewState extends State<ArCoreFaceView>
     with WidgetsBindingObserver {
   @override
   void initState() {
-    WidgetsBinding.instance.addObserver(this);
+    WidgetsBinding.instance!.addObserver(this);
     super.initState();
   }
 
@@ -53,7 +53,7 @@ class _ArCoreFaceViewState extends State<ArCoreFaceView>
     if (widget.onArCoreViewCreated == null) {
       return;
     }
-    widget.onArCoreViewCreated(
+    widget.onArCoreViewCreated!(
       ArCoreFaceController(
         id: id,
         enableAugmentedFaces: widget.enableAugmentedFaces,
@@ -63,7 +63,7 @@ class _ArCoreFaceViewState extends State<ArCoreFaceView>
 
   @override
   void dispose() {
-    WidgetsBinding.instance.removeObserver(this);
+    WidgetsBinding.instance!.removeObserver(this);
     super.dispose();
   }
 }
