@@ -2,15 +2,15 @@ import 'package:vector_math/vector_math_64.dart';
 import 'arcore_pose.dart';
 
 class ArCoreHitTestResult {
-  double distance;
+  double? distance;
 
-  Vector3 translation;
+  Vector3? translation;
 
-  Vector4 rotation;
+  Vector4? rotation;
 
-  String nodeName;
+  String? nodeName;
 
-  ArCorePose pose;
+  ArCorePose? pose;
 
   ArCoreHitTestResult.fromMap(Map<dynamic, dynamic> map) {
     this.distance = map['distance'];
