@@ -10,10 +10,10 @@ class ArCoreNode {
   ArCoreNode({
     this.shape,
     this.image,
-    String name,
-    Vector3 position,
-    Vector3 scale,
-    Vector4 rotation,
+    String? name,
+    Vector3? position,
+    Vector3? scale,
+    Vector4? rotation,
     this.children = const [],
   })  : name = name ?? random_string.randomString(),
         position = ValueNotifier(position),
@@ -23,17 +23,17 @@ class ArCoreNode {
 
   final List<ArCoreNode> children;
 
-  final ArCoreShape shape;
+  final ArCoreShape? shape;
 
-  final ValueNotifier<Vector3> position;
+  final ValueNotifier<Vector3?> position;
 
-  final ValueNotifier<Vector3> scale;
+  final ValueNotifier<Vector3?> scale;
 
-  final ValueNotifier<Vector4> rotation;
+  final ValueNotifier<Vector4?> rotation;
 
   final String name;
 
-  final ArCoreImage image;
+  final ArCoreImage? image;
 
   Map<String, dynamic> toMap() => <String, dynamic>{
         'dartType': runtimeType.toString(),
