@@ -7,11 +7,11 @@ import 'package:arcore_flutter_plugin/src/shape/arcore_shape.dart';
 class ArCoreRotatingNode extends ArCoreNode {
   ArCoreRotatingNode({
     this.shape,
-    double degreesPerSecond,
-    Vector3 position,
-    Vector3 scale,
-    Vector4 rotation,
-    String name,
+    double? degreesPerSecond,
+    Vector3? position,
+    Vector3? scale,
+    Vector4? rotation,
+    String? name,
   })  : degreesPerSecond = ValueNotifier(90.0),
         super(
           shape: shape,
@@ -20,7 +20,7 @@ class ArCoreRotatingNode extends ArCoreNode {
           scale: scale,
         );
 
-  final ArCoreShape shape;
+  final ArCoreShape? shape;
 
   final ValueNotifier<double> degreesPerSecond;
 
