@@ -7,12 +7,12 @@ import 'package:arcore_flutter_plugin/src/shape/arcore_shape.dart';
 class ArCoreCube extends ArCoreShape {
   ArCoreCube({
     this.size,
-    List<ArCoreMaterial> materials,
+    List<ArCoreMaterial>? materials,
   }) : super(
           materials: materials,
         );
 
-  final Vector3 size;
+  final Vector3? size;
 
   @override
   Map<String, dynamic> toMap() => <String, dynamic>{
