@@ -4,7 +4,7 @@ import 'package:arcore_flutter_plugin/src/shape/arcore_shape.dart';
 class ArCoreSphere extends ArCoreShape {
   ArCoreSphere({
     this.radius = 0.5,
-    List<ArCoreMaterial> materials,
+    List<ArCoreMaterial>? materials,
   }) : super(
           materials: materials,
         );
